+++ .vscode/launch.json
@@ -1,0 +1,20 @@
+{
+    // Use IntelliSense to learn about possible attributes.
+    // Hover to view descriptions of existing attributes.
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "name": "Python: Cifar",
+            "type": "python",
+            "request": "launch",
+            "program": "${workspaceFolder}/augmix_refactored/script/cifar.py",
+            "console": "integratedTerminal",
+            "justMyCode": true,
+            "args": [
+                "--config-path",
+                "${workspaceFolder}/cifar_config.yaml"
+            ]
+        }
+    ]
+}
