+++ cifar_config.yaml
@@ -1,0 +1,23 @@
+Config:
+  all_ops: false
+  augmentation_severity: 3
+  batch_size: 128
+  dataset: cifar10
+  image_size: 32
+  droprate: 0.1
+  epochs: 5
+  eval_batch_size: 1000
+  evaluate: false
+  layers: 20
+  learning_rate: 0.1
+  mixture_depth: 4
+  mixture_width: 3
+  model: wrn
+  momentum: 0.9
+  no_jsd: false
+  num_workers: 1
+  print_freq: 50
+  resume_path: ''
+  save_folder: ./snapshots
+  weight_decay: 0.0005
+  widen_factor: 2
