+++ docs/howto.md
@@ -1,0 +1,3 @@
+# HowTo Augmix
+
+This should hold some explanation...
