--- requirements.txt
@@ -1,4 +1,0 @@
-numpy>=1.15.0
-Pillow>=6.1.0
-torch==1.2.0
-torchvision==0.2.2
