+++ snapshots/cifar10_wrn_training_log.csv
@@ -1,0 +1,3 @@
+epoch,time(s),train_loss,test_loss,test_error(%)
+001,00098,1.266393,0.00111,40.31
+002,00089,0.984167,0.00078,26.51
